AP-41 Alterado todos os alerts pelo sweetAlert

+++ src/webapp/assets/js/publicacoes.js
@@ -44,7 +44,11 @@ function curtirPublicacao(evento) {
         elementoClicado.removeClass('curtir-publicacao');
 
     }).fail(function(){
-        alert("Erro ao curtir publicação");
+        Swal.fire(
+            'Ops...',
+            'Erro ao criar publicação!',
+            'error'
+        );
         console.log(publicacaoId);
     }).always(function() {
         elementoClicado.prop('disabled', false);
@@ -72,7 +76,11 @@ function descurtirPublicacao(evento) {
         elementoClicado.addClass('curtir-publicacao');
 
     }).fail(function(){
-        alert("Erro ao curtir publicação");
+        Swal.fire(
+            'Ops...',
+            'Erro ao curtir publicação!',
+            'error'
+        );
         console.log(publicacaoId);
     }).always(function() {
         elementoClicado.prop('disabled', false);
@@ -92,9 +100,20 @@ function atualizarPublicacao() {
             conteudo: $('#conteudo').val()
         }
     }).done(function(){
-        alert("Publicação editada com sucesso");
+        Swal.fire(
+            'Sucesso!',
+            'Publicação Atualizada com sucesso!',
+            'success'
+        )
+        .then(function(){
+            window.location = "/web/home";
+        })
     }).fail(function(){
-        alert("Erro ao editar a publicação");
+        Swal.fire(
+            'Ops...',
+            'Erro ao editar publicação!',
+            'error'
+        );;
     }).always(function(){
         $('#atualizar-publicacao').prop('disabled', false);
     })
@@ -103,20 +122,35 @@ function atualizarPublicacao() {
 function deletarPublicacao(evento) {
     evento.preventDefault();
 
-    const elementoClicado = $(evento.target);
-    const publicacao = elementoClicado.closest('div')
-    const publicacaoId = publicacao.data('publicacao-id');
+    Swal.fire({
+        title: "Atenção!",
+        text: "Tem certeza que deseja excluir essa publicação? Essa ação é irreversível!",
+        showCancelButton: true,
+        cancelButtonText: "Cancelar",
+        icon: "warning"
+    }).then(function(confirmacao){
+        if (!confirmacao.value) return;
 
-    elementoClicado.prop('disabled', true);
-
-    $.ajax({
-        url: `/web/publicacoes/${publicacaoId}/deletar`,
-        method: "DELETE"
-    }).done(function(){
-        publicacao.fadeOut("slow", function(){
-            $(this).remove();
+        const elementoClicado = $(evento.target);
+        const publicacao = elementoClicado.closest('div')
+        const publicacaoId = publicacao.data('publicacao-id');
+    
+        elementoClicado.prop('disabled', true);
+    
+        $.ajax({
+            url: `/web/publicacoes/${publicacaoId}/deletar`,
+            method: "DELETE"
+        }).done(function(){
+            publicacao.fadeOut("slow", function(){
+                $(this).remove();
+            });
+        }).fail(function(){
+            Swal.fire(
+                'Ops...',
+                'Erro ao excluir publicação!',
+                'error'
+            );
         });
-    }).fail(function(){
-        alert("Erro ao excluir a publicação");
-    });
+    })
+
 }
